Refactor FavouriteButton
