+++ src/components/favourite-launch-button.js
@@ -1,23 +1,14 @@
 import React from "react";
-import { Star } from "react-feather";
-import { IconButton } from "@chakra-ui/core";
 import { useFavouriteLaunch } from "../utils/favourites-context";
+import FavouriteButton from "./favourite-button";
 
 export default function FavouriteLaunchButton({ launchId }) {
   const { isFavourite, toggleFavouriteLaunch } = useFavouriteLaunch(launchId);
 
-  function handleFavouriteClick(event) {
-    event.preventDefault();
-    event.stopPropagation();
-    toggleFavouriteLaunch();
-  }
-
   return (
-    <IconButton
-      aria-label="Favourite button"
-      variantColor={isFavourite ? "teal" : undefined}
-      icon={Star}
-      onClick={handleFavouriteClick}
+    <FavouriteButton
+      isFavourite={isFavourite}
+      onClick={toggleFavouriteLaunch}
     />
   );
 }
