+++ src/components/favourite-launch-pad-button.js
@@ -1,23 +1,14 @@
 import React from "react";
-import { Star } from "react-feather";
-import { IconButton } from "@chakra-ui/core";
 import { useFavouriteLaunchPad } from "../utils/favourites-context";
+import FavouriteButton from "./favourite-button";
 
 export default function FavouriteLaunchPadButton({ launchPadId }) {
   const { isFavourite, toggleFavouriteLaunchPad } = useFavouriteLaunchPad(launchPadId);
 
-  function handleFavouriteClick(event) {
-    event.preventDefault();
-    event.stopPropagation();
-    toggleFavouriteLaunchPad();
-  }
-
   return (
-    <IconButton
-      aria-label="Favourite button"
-      variantColor={isFavourite ? "teal" : undefined}
-      icon={Star}
-      onClick={handleFavouriteClick}
+    <FavouriteButton
+      isFavourite={isFavourite}
+      onClick={toggleFavouriteLaunchPad}
     />
   );
 }
